backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "fastapi>=0.121.2",
    "html5lib>=1.1",
    "langchain>=1.0.7",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.3",
    "langchain-qdrant>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "lxml>=6.0.2",
    "openai>=2.8.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.15.1",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]
